+++ src/core/nyx.cpp
@@ -37,11 +37,12 @@
 #include "core/nyx.h"
 
 // clang-format off
-#include <CL/cl_gl.h>
 #include <GL/glew.h>
+#define SDL_MAIN_HANDLED
 #include <SDL2/SDL.h>
 #include <SDL2/SDL_opengl.h>
 #include <GL/gl.h>
+#include <CL/cl_gl.h>
 // clang-format on
 
 #include "compute/compute_cpu.h"
@@ -55,7 +56,6 @@
 #include "gui/rgb_triangle.h"
 #include "gui/vao_triangle.h"
 #include "gui/gl_matrices.h"
-#include "gui/interop.h"
 #include "gui/cl_particles.h"
 #include "gui/rgb_cube.h"
 #include "gui/rgb_cube_texture.h"
@@ -322,12 +322,6 @@ int main(int argc, char *argv[])
                     glm.run();
                     break;
                 }
-                case 6:
-                {
-                    interop iop;
-                    iop.run();
-                    break;
-                }
                 case 7:
                 {
                     cl_particles clp;
@@ -382,7 +376,6 @@ void print_help()
     std::cout << "                                      4 - draw OpenGL VAO triangle" << std::endl;
     std::cout << "                                      5 - draw OpenGL VAO triangle with matrices support" << std::endl;
     std::cout << "                                      6 - draw Mandelbrot set via OpenGL with interoperability" << std::endl;
-    std::cout << "                                      7 - draw OpenCL particles with interoperability (don't work)" << std::endl;
     std::cout << "                                      8 - draw OpenGL RGB cube" << std::endl;
     std::cout << "                                      9 - draw OpenGL RGB textured cube" << std::endl;
     std::cout << "  -b, --verbose                   Verbose output" << std::endl;