+++ src/gui/gl_matrices.cpp
@@ -41,7 +41,8 @@
 #include <string>
 #include <type_traits>
 
-#include "SDL_keycode.h"
+#include <SDL2/SDL_keycode.h>
+
 #include "gui/program.h"
 #include "io/log/logger.h"